pyproject.toml
[tool.poetry]
name = "bot_strategy"
version = "0.1.0"
description = ""
authors = ["Max Ermakov <maksim_ermakov@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.17.0"
tinvest = "^3.0.3"
asyncio = "^3.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
